Sort完成

+++ src/main/java/cn/Bookspf/model/DO/DBOrder.java
@@ -5,72 +5,90 @@ import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 public class DBOrder {
-	private  long id;
-	private long orderid;
-	private int uid;
-	private int bid;
+	private Integer id;
+	private Integer orderid;
+	private Integer uid;
+	private Integer bid;
 	private String isbn;
 	private Double bookprice;
-	private int paid;
+	private Integer paid;
 	@JsonFormat(pattern="yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
 	private Date createtime;
 	@JsonFormat(pattern="yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
 	private Date paytime;
-	public long getId() {
+	
+	public Integer getId() {
 		return id;
 	}
-	public void setId(long id) {
+
+	public void setId(Integer id) {
 		this.id = id;
 	}
-	public long getOrderid() {
+
+	public Integer getOrderid() {
 		return orderid;
 	}
-	public void setOrderid(long orderid) {
+
+	public void setOrderid(Integer orderid) {
 		this.orderid = orderid;
 	}
-	public int getUid() {
+
+	public Integer getUid() {
 		return uid;
 	}
-	public void setUid(int uid) {
+
+	public void setUid(Integer uid) {
 		this.uid = uid;
 	}
-	public int getBid() {
+
+	public Integer getBid() {
 		return bid;
 	}
-	public void setBid(int bid) {
+
+	public void setBid(Integer bid) {
 		this.bid = bid;
 	}
+
 	public String getIsbn() {
 		return isbn;
 	}
+
 	public void setIsbn(String isbn) {
 		this.isbn = isbn;
 	}
+
 	public Double getBookprice() {
 		return bookprice;
 	}
+
 	public void setBookprice(Double bookprice) {
 		this.bookprice = bookprice;
 	}
-	public int getPaid() {
+
+	public Integer getPaid() {
 		return paid;
 	}
-	public void setPaid(int paid) {
+
+	public void setPaid(Integer paid) {
 		this.paid = paid;
 	}
+
 	public Date getCreatetime() {
 		return createtime;
 	}
+
 	public void setCreatetime(Date createtime) {
 		this.createtime = createtime;
 	}
+
 	public Date getPaytime() {
 		return paytime;
 	}
+
 	public void setPaytime(Date paytime) {
 		this.paytime = paytime;
 	}
-	
+
 	@Override
 	public String toString() {
 		return "DBOrder [id=" + id + ", orderid=" + orderid + ", uid=" + uid + ", bid=" + bid + ", isbn=" + isbn
